pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Meet Mangukiya <meetmangukiya98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
aiocron = "^1.3"
aiofiles = "^0.5.0"
motor = "^2.1.0"
aioinflux = "^0.9.0"
protobuf = "^3.12.2"
pandas = "^1.0.4"
pytest = "^5.4.3"
pandas-market-calendars = "^1.4.2"
click = "^7.1.2"
pyyaml = "^5.3.1"
aioredis = "^1.3.1"
python-dotenv = "^0.15.0"
lxml = "^4.6.2"

[tool.poetry.dev-dependencies]
aiohttp-devtools = "^0.13.1"
black = "^20.8b1"
pyflakes = "^2.2.0"
isort = "^4.3.21"
pytest = "^5.4.3"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
